requirements.txt
numpy
ipykernel
pandas
seaborn
matplotlib
mlflow
pyspark
dash
xgboost
scikit-learn
python-dotenv
dash_daq